Tools for accessing the fields of a record.

A rcrd behaves like a vector, so length(), names(), and $ can not provide access to the fields of the underlying list. These helpers do: fields() is equivalent to names(); n_fields() is equivalent to length(); field() is equivalent to $.

Usage

fields(x)

n_fields(x)

field(x, i)

field(x, i) <- value

Arguments

x

A rcrd, i.e. a list of equal length vectors with unique names.

Examples

x <- new_rcrd(list(x = 1:3, y = 3:1, z = letters[1:3]))
n_fields(x)
#> [1] 3
fields(x)
#> [1] "x" "y" "z"

field(x, "y")
#> [1] 3 2 1
field(x, "y") <- runif(3)
field(x, "y")
#> [1] 0.5335507 0.2900965 0.6150181