new_data_frame {vctrs}	R Documentation

Data frame class

Description

A data.frame data.frame() is a list with "row.names" attribute. Each element of the list must be named, and of the same length. These functions help the base data.frame classes fit in to the vctrs type system by providing constructors, coercion functions, and casting functions.

Usage

new_data_frame(x = list(), n = NULL, ..., class = character())

## S3 method for class 'data.frame'
vec_ptype2(x, y, ...)

## S3 method for class 'data.frame'
vec_cast(x, to, ...)

Arguments

x
A named list of equal-length vectors. The lengths are not checked; it is responsibility of the caller to make sure they are equal.

n
Number of rows. If NULL, will be computed from the length of the first element of x.

..., class
Additional arguments for creating subclasses.

Examples

new_data_frame(list(x = 1:10, y = 10:1))

[Package vctrs version 0.2.0 Index]