new_rcrd {vctrs}	R Documentation

rcrd (record) S3 class

Description

The rcrd class extends vctr. A rcrd is composed of 1 or more fields, which must be vectors of the same length. Is designed specifically for classes that can naturally be decomposed into multiple vectors of the same length, like POSIXlt, but where the organisation should be considered an implementation detail invisible to the user (unlike a data.frame).

Usage

new_rcrd(fields, ..., class = character())

Arguments

fields
A list or a data frame. Lists must be rectangular (same sizes), and contain uniquely named vectors (at least one). fields is validated with df_list() to ensure uniquely named vectors.

...
Additional attributes

class
Name of subclass.

[Package vctrs version 0.6.4 Index]