vctr (vectr) S3 class

This abstract class provides a set of useful default methods that makes it considerably easier to get started with a new S3 vector class.

new_vctr(.data, ..., class = character())

Arguments

...
Name-value pairs defining attributes

class
Name of subclass.

x
Foundation of class. Must be a vector

Recommended workflow

Start by creating a low-level constructor. It should be called new_myclass(), and should check the types (but not the values) of its inputs.

Depending on the class create either a helper myclass(), a coercer as_myclass(), or both. A helper should construct valid values then pass on to constructor. A coercer should either check values are correct, or rely on the helper. Avoid defining validation code in multiple places.

Define a useful format.myclass() method. This will give default print() and as.character() methods that should be adequate for most classes. Be warned: a good format method may be as much work as the rest of the class put together!

Next provide vec_type2() and vec_cast(). First focus on the casts between your class and its underlying the base type. Next think about base types that should be coercible or castable. See vignette("s3-vector") for details.

If your function behaves similarly to numbers or booleans, or has specialised comparison methods, read vec_grp to learn about the vctrs group generics: these allow you to implement many methods at once.

Implementing these methods gets you many methods for free:

[[ and [ use NextMethod() dispatch to the underlying base function, reconstructing attributes with vec_cast(). rep() works similarly. Override if one or more attributes have a one-to-one relationship to the underlying data.

[[<- and [<- cast the RHS to the LHS, then call NextMethod(). Override these methods if any attributes depend on the data.

as.list.vctrs_vctr() calls [[ repeatedly, as.character.vctrs_vctr() calls format().

as.data.frame.vctrs_vctr() uses a standard technique to wrap a vector in a data frame. You should never need to override this method.

dims<-.vctrs_vctr(), and dimnames<-.vctrs_vctr() all throw errors as generally custom vector classes do not need to support dimensions.