partial_frame {vctrs}	R Documentation

Partially specify columns of a data frame

Description

This special class can be passed to .ptype in order to specify the types of only some of the columns in a data frame.

Usage

partial_frame(...)

Arguments

...
Attributes of subclass

Examples

pf <- partial_frame(x = double())
pf

vec_rbind(
  data.frame(x = 1L, y = "a"),
  data.frame(x = FALSE, z = 10),
  .ptype = partial_frame(x = double(), a = character())
)

[Package vctrs version 0.6.4 Index]