table {vctrs}	R Documentation

Table S3 class

Description

These functions help the base table class fit into the vctrs type system by providing coercion and casting functions.

[Package vctrs version 0.5.0 Index]