unspecified {vctrs}	R Documentation

A 1d vector of unspecified type

Description

This is a partial type used to represent logical vectors that only contain NA. These require special handling because we want to allow NA to specify missingness without requiring a type.

Usage

unspecified(n = 0)

Arguments

n
Length of vector

Examples

vec_ptype_show()
vec_ptype_show(NA)

vec_c(NA, factor("x"))
vec_c(NA, Sys.Date())
vec_c(NA, Sys.time())
vec_c(NA, list(1:3, 4:5))

[Package vctrs version 0.6.4 Index]