vctrs-package {vctrs}	R Documentation

vctrs: Vector Helpers

Description

Defines new notions of prototype and size that are used to provide tools for consistent and well-founded type-coercion and size-recycling, and are in turn connected to ideas of type- and size-stability useful for analysing function interfaces.

Author(s)

Maintainer: Davis Vaughan davis@posit.co

Authors:

Hadley Wickham hadley@posit.co

Lionel Henry lionel@posit.co

Other contributors:

data.table team (Radix sort based on data.table's forder() and their contribution to R's order()) [copyright holder]

Posit Software, PBC [copyright holder, funder]

See Also

Useful links:

https://vctrs.r-lib.org/

https://github.com/r-lib/vctrs

Report bugs at https://github.com/r-lib/vctrs/issues

[Package vctrs version 0.6.5 Index]