Repair names with legacy method

This standardises names with the legacy approach that was used in tidyverse packages (such as tibble, tidyr, and readxl) before vec_as_names() was implemented. This tool is meant to help transitioning to the new name repairing standard and will be deprecated and removed from the package some time in the future.

vec_as_names_legacy(names, prefix = "V", sep = "")

Arguments

names
A character vector.

prefix, sep
Prefix and separator for repaired names.

Examples

if (rlang::is_installed("tibble")) {

library(tibble)

# Names repair is turned off by default in tibble:
try(tibble(a = 1, a = 2))

# You can turn it on by supplying a repair method:
tibble(a = 1, a = 2, .name_repair = "universal")

# If you prefer the legacy method, use `vec_as_names_legacy()`:
tibble(a = 1, a = 2, .name_repair = vec_as_names_legacy)

}

#> Error : Column name `a` must not be duplicated.
#> Use .name_repair to specify repair.
#> Backtrace:
#>      █
#>   1. ├─base::tryCatch(...)
#>   2. │ └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>   3. │   ├─base:::tryCatchOne(...)
#>   4. │   │ └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   5. │   └─base:::tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#>   6. │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>   7. │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   8. ├─base::withCallingHandlers(...)
#>   9. ├─base::saveRDS(...)
#>  10. ├─base::do.call(...)
#>  11. ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#>  12. └─(function (..., crayon_enabled, crayon_colors, pkgdown_internet) ...
#>  13.   └─pkgdown::build_site(...)
#>  14.     └─pkgdown:::build_site_local(...)
#>  15.       └─pkgdown::build_reference(...)
#>  16.         └─purrr::map(...)
#>  17.           └─pkgdown:::.f(.x[[i]], ...)
#>  18.             └─pkgdown:::data_reference_topic(...)
#>  19.               ├─pkgdown:::as_data(...)
#>  20.               └─pkgdown:::as_data.tag_examples(...)
#>  21.                 └─purrr::pmap_chr(...)
#>  22.                   └─pkgdown:::.f(...)
#>  23.                     ├─withr::with_options(...)
#>  24.                     │ └─base::force(code)
#>  25.                     └─evaluate::evaluate(code, env, new_device = TRUE)
#>  26.                       └─evaluate:::evaluate_call(...)
#>  27.                         ├─evaluate:::timing_fn(...)
#>  28.                         ├─evaluate:::handle(...)
#>  29.                         │ └─base::try(f, silent = TRUE)
#>  30.                         │   └─base::tryCatch(...)
#>  31.                         │     └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  32.                         │       └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  33.                         │         └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  34.                         ├─base::withCallingHandlers(...)
#>  35.                         ├─base::withVisible(eval(expr, envir, enclos))
#>  36.                         └─base::eval(expr, envir, enclos)
#>  37.                           └─base::eval(expr, envir, enclos)
#>  38.                             ├─base::try(tibble(a = 1, a = 2))
#>  39.                             │ └─base::tryCatch(...)
#>  40.                             │   └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  41.                             │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  42.                             │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  43.                             └─tibble::tibble(a = 1, a = 2)
#>  44.                               └─tibble:::lst_to_tibble(xlq$output, .rows, .name_repair, lengths = xlq$lengths)
#>  45.                                 └─tibble:::set_repaired_names(x, .name_repair)
#>  46.                                   ├─rlang::set_names(x, repaired_names(names(x), .name_repair = .name_repair))
#>  47.                                   │ └─rlang:::set_names_impl(x, x, nm, ...)
#>  48.                                   │   └─rlang::is_function(nm)
#>  49.                                   │     └─rlang::is_closure(x)
#>  50.                                   └─tibble:::repaired_names(names(x), .name_repair = .name_repair)
#>  51.                                     └─tibble:::check_unique(new_name)

#> New names:
#> * a -> a...1
#> * a -> a...2

#> # A tibble: 1 x 2
#>       a    a1
#>   <dbl> <dbl>
#> 1     1     2