Coerce vectors to common type

This function is primarily useful as a building block for other functions. The implementation also illustrates the key ideas of the vctrs type system, and allows you to experiment interactively.

vec_coerce(..., .ptype = NULL)

Arguments

...
Vectors to coerce.

.ptype
If NULL, the default, the output type is determined by computing the common type across all elements of ....

Alternatively, you can supply .ptype to give the output known type. If getOption("vctrs.no_guessing") is TRUE you must supply this value: this is a convenient way to make production code demand fixed types.

Value

A list_of input vectors coerced to .ptype, an error stating that a common type could not be found, or an error stating that casting a input to .ptype was not possible.

Examples

vec_coerce(factor("a"), factor(c("a", "b")))

#> <list_of<factor<078e5>>[2]>
#> [[1]]
#> [1] a
#> Levels: a b
#> 
#> [[2]]
#> [1] a b
#> Levels: a b
#> 

vec_coerce(factor("a"), Sys.Date(), .ptype = list())

#> <list_of<list>[2]>
#> [[1]]
#> [[1]][[1]]
#> [1] a
#> Levels: a
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> [1] "2018-09-12"
#> 
#>