Order vectors with base compatibility

vec_order_base() orders vectors using base::order(), but can handle more complex types, like data frames and vctrs_vctr objects, using vctrs principles.

vec_order_base() is mainly provided for backwards compatibility with vctrs <= 0.3.7. New code should instead use vec_order(), which has more capabilities. The main difference between the two is that vec_order() orders character vectors in the C locale (which is highly performant), while vec_order_base() respects the system locale.

vec_order_base(
  x,
  direction = c("asc", "desc"),
  na_value = c("largest", "smallest")
)

Arguments

x

A vector

direction

Direction to sort in. Defaults to ascending.

na_value

Should NAs be treated as the largest or smallest values?

Value

An integer vector the same size as x.

Differences with order()

Unlike the na.last argument of order() which decides the positions of missing values irrespective of the decreasing argument, the na_value argument of vec_order_base() interacts with direction. If missing values are considered the largest value, they will appear last in ascending order, and first in descending order.

Dependencies of vec_order_base()

vec_proxy_order()

Examples

x <- round(c(runif(9), NA), 3)
vec_order_base(x)
#>  [1]  2  7  9  4  8  3  5  1  6 10
vec_order_base(x, "desc")
#>  [1] 10  6  1  5  3  8  4  9  7  2

# Can also handle data frames
df <- data.frame(g = sample(2, 10, replace = TRUE), x = x)
vec_order_base(df)
#>  [1]  2  9  4  8  3  1  6 10  7  5
vec_order_base(df, "desc")
#>  [1]  5  7 10  6  1  3  8  4  9  2

# Missing values interpreted as largest values are last when
# in increasing order:
vec_order_base(c(1, NA), na_value = "largest", direction = "asc")
#> [1] 1 2
vec_order_base(c(1, NA), na_value = "largest", direction = "desc")
#> [1] 2 1