Find the prototype of a set of vectors

This function is aimed at developers, but may also be useful for users who want to explore the type system interactively. Use this function to find the common type from a set of inputs.

vec_ptype(..., .ptype = NULL)

Arguments

...
Vectors inputs

.ptype
If NULL, the default, the output type is determined by computing the common type across all elements of ....

Alternatively, you can supply .ptype to give the output known type. If getOption("vctrs.no_guessing") is TRUE you must supply this value: this is a convenient way to make production code demand fixed types.

Value

A prototype

Details

This function works by finding the prototype (a zero-length subset) of each input, then using Reduce() and vec_type2() to find the common class. NULLs and logical vectors that consist only of NA (including those inside data frames) are converted to the special unknown type. This is needed because bare NAs should be automatically coercible to every other class.

Examples

# Unknown types ------------------------------------------
vec_ptype()

#> prototype: unknown

vec_ptype(NA)

#> prototype: unknown

vec_ptype(NULL)

#> prototype: unknown

# Vectors ------------------------------------------------
vec_ptype(1:10)

#> prototype: integer

vec_ptype(letters)

#> prototype: character

vec_ptype(TRUE)

#> prototype: logical

vec_ptype(Sys.Date())

#> prototype: date

vec_ptype(Sys.time())

#> prototype: datetime<local>

vec_ptype(factor("a"))

#> prototype: factor<127a2>

vec_ptype(ordered("a"))

#> prototype: ordered<127a2>

# Matrices -----------------------------------------------
# The prototype of a matrix includes the number of columns
vec_ptype(array(1, dim = c(1, 2)))

#> prototype: double[,2]

vec_ptype(array("x", dim = c(1, 2)))

#> prototype: character[,2]

# Data frames --------------------------------------------
# The prototype of a data frame includes the prototype of
# every column
vec_ptype(iris)

#> prototype: data.frame<
#>   Sepal.Length: double
#>   Sepal.Width : double
#>   Petal.Length: double
#>   Petal.Width : double
#>   Species     : factor<12d60>
#> >

# The prototype of multiple data frames includes the prototype
# of every column that in any data frame
vec_ptype(
  data.frame(x = TRUE),
  data.frame(y = 2),
  data.frame(z = "a")
)

#> prototype: data.frame<
#>   x: logical
#>   y: double
#>   z: factor<127a2>
#> >