vec_ptype_full {vctrs}	R Documentation

Vector type as a string

Description

vec_ptype_full() displays the full type of the vector. vec_ptype_abbr() provides an abbreviated summary suitable for use in a column heading.

Usage

vec_ptype_full(x, ...)

vec_ptype_abbr(x, ...)

Arguments

x
A vector.

...
These dots are for future extensions and must be empty.

Value

A string.

S3 dispatch

The default method for vec_ptype_full() uses the first element of the class vector. Override this method if your class has parameters that should be prominently displayed.

The default method for vec_ptype_abbr() abbreviate()s vec_ptype_full() to 8 characters. You should almost always override, aiming for 4-6 characters where possible.

Examples

cat(vec_ptype_full(1:10))
cat(vec_ptype_full(iris))

cat(vec_ptype_abbr(1:10))

[Package vctrs version 0.2.0 Index]