new_partial {vctrs}	R Documentation

Partial type

Description

Use new_partial() when constructing a new partial type subclass; and use is_partial() to test if a type is partial. All subclasses need to provide a vec_ptype_finalise() method.

Usage

new_partial(..., class = character())

is_partial(x)

vec_ptype_finalise(x, ...)

Arguments

...
Attributes of the partial type

class
Name of subclass.

Details

As the name suggests, a partial type partially specifies a type, and it must be combined with data to yield a full type. A useful example of a partial type is partial_frame(), which makes it possible to specify the type of just a few columns in a data frame. Use this constructor if you're making your own partial type.

[Package vctrs version 0.6.4 Index]