rcrd (record) S3 class

The rcrd class extends vctr. A rcrd is composed of 1 or more fields, which must be vectors of the same length. Is designed specifically for classes that can naturally be decomposed into multiple vectors of the same length, like POSIXlt, but where the organisation should be considered an implementation detail invisible to the user (unlike a data.frame).

Usage

new_rcrd(fields, ..., class = character())

Arguments

fields

A list or a data frame. Lists must be rectangular (same sizes), and contain uniquely named vectors (at least one). fields is validated with df_list() to ensure uniquely named vectors.

...

Additional attributes

class

Name of subclass.

Details

Record-style objects created with new_rcrd() do not do much on their own. For instance they do not have a default format() method, which means printing the object causes an error. See Record-style objects for details on implementing methods for record vectors.