A 1d vector of unspecified type

This is a partial type used to represent logical vectors that only contain NA. These require special handling because we want to allow NA to specify missingness without requiring a type.

unspecified(n = 0)

Arguments

n
Length of vector

Examples

vec_ptype_show()

#> Prototype: NULL

vec_ptype_show(NA)

#> Prototype: logical

vec_c(NA, factor("x"))

#> [1] <NA> x   
#> Levels: x

vec_c(NA, Sys.Date())

#> [1] NA           "2020-08-27"

vec_c(NA, Sys.time())

#> [1] NA                        "2020-08-27 08:56:00 UTC"

vec_c(NA, list(1:3, 4:5))

#> [[1]]
#> NULL
#> 
#> [[2]]
#> [1] 1 2 3
#> 
#> [[3]]
#> [1] 4 5
#>