Count unique values in a vector

Count the number of unique values in a vector. vec_count() has two important differences to table(): it returns a data frame, and when given multiple inputs (as a data frame), it only counts combinations that appear in the input.

Usage

vec_count(x, sort = c("count", "key", "location", "none"))

Arguments

x

A vector (including a data frame).

sort

One of "count", "key", "location", or "none".

"count", the default, puts most frequent values at top

"key", orders by the output key column (i.e. unique values of x)

"location", orders by location where key first seen. This is useful if you want to match the counts up to other unique/duplicated functions.

"none", leaves unordered. This is not guaranteed to produce the same ordering across R sessions, but is the fastest method.

Value

A data frame with columns key (same type as x) and count (an integer vector).

Dependencies

vec_proxy_equal()

vec_slice()

vec_order()

Examples

vec_count(mtcars$vs)
#>   key count
#> 1   0    18
#> 2   1    14
vec_count(iris$Species)
#>          key count
#> 1     setosa    50
#> 2 versicolor    50
#> 3  virginica    50

# If you count a data frame you'll get a data frame
# column in the output
str(vec_count(mtcars[c("vs", "am")]))
#> 'data.frame':	4 obs. of  2 variables:
#>  $ key  :'data.frame':	4 obs. of  2 variables:
#>   ..$ vs: num  0 1 1 0
#>   ..$ am: num  0 1 0 1
#>  $ count: int  12 7 7 6

# Sorting ---------------------------------------

x <- letters[rpois(100, 6)]
# default is to sort by frequency
vec_count(x)
#>    key count
#> 1    g    18
#> 2    e    17
#> 3    f    16
#> 4    d    14
#> 5    i     8
#> 6    c     7
#> 7    h     6
#> 8    b     5
#> 9    k     5
#> 10   j     3
#> 11   l     1

# by can sort by key
vec_count(x, sort = "key")
#>    key count
#> 1    b     5
#> 2    c     7
#> 3    d    14
#> 4    e    17
#> 5    f    16
#> 6    g    18
#> 7    h     6
#> 8    i     8
#> 9    j     3
#> 10   k     5
#> 11   l     1

# or location of first value
vec_count(x, sort = "location")
#>    key count
#> 1    e    17
#> 2    f    16
#> 3    h     6
#> 4    i     8
#> 5    j     3
#> 6    d    14
#> 7    g    18
#> 8    c     7
#> 9    b     5
#> 10   k     5
#> 11   l     1
head(x)
#> [1] "e" "f" "h" "f" "i" "e"

# or not at all
vec_count(x, sort = "none")
#>    key count
#> 1    k     5
#> 2    c     7
#> 3    h     6
#> 4    e    17
#> 5    l     1
#> 6    j     3
#> 7    g    18
#> 8    f    16
#> 9    d    14
#> 10   i     8
#> 11   b     5