Order and sort vectors

Order and sort vectors

vec_order(x, direction = c("asc", "desc"), na_value = c("largest", "smallest"))

vec_sort(x, direction = c("asc", "desc"), na_value = c("largest", "smallest"))

Arguments

x

A vector

direction

Direction to sort in. Defaults to ascending.

na_value

Should NAs be treated as the largest or smallest values?

Value

vec_order() an integer vector the same size as x.

vec_sort() a vector with the same size and type as x.

Differences with order()

Unlike the na.last argument of order() which decides the positions of missing values irrespective of the decreasing argument, the na_value argument of vec_order() interacts with direction. If missing values are considered the largest value, they will appear last in ascending order, and first in descending order.

Dependencies of vec_order()

vec_proxy_order()

Dependencies of vec_sort()

vec_proxy_order()

vec_order()

vec_slice()

Examples

x <- round(c(runif(9), NA), 3)
vec_order(x)
#>  [1]  8  2  5  1  4  9  7  3  6 10
vec_sort(x)
#>  [1] 0.134 0.180 0.251 0.584 0.654 0.765 0.817 0.855 0.943    NA
vec_sort(x, "desc")
#>  [1]    NA 0.943 0.855 0.817 0.765 0.654 0.584 0.251 0.180 0.134

# Can also handle data frames
df <- data.frame(g = sample(2, 10, replace = TRUE), x = x)
vec_order(df)
#>  [1]  8  2  1  9  5  4  7  3  6 10
vec_sort(df)
#>    g     x
#> 1  1 0.134
#> 2  1 0.180
#> 3  1 0.584
#> 4  1 0.765
#> 5  2 0.251
#> 6  2 0.654
#> 7  2 0.817
#> 8  2 0.855
#> 9  2 0.943
#> 10 2    NA
vec_sort(df, "desc")
#>    g     x
#> 1  2    NA
#> 2  2 0.943
#> 3  2 0.855
#> 4  2 0.817
#> 5  2 0.654
#> 6  2 0.251
#> 7  1 0.765
#> 8  1 0.584
#> 9  1 0.180
#> 10 1 0.134

# Missing values interpreted as largest values are last when
# in increasing order:
vec_order(c(1, NA), na_value = "largest", direction = "asc")
#> [1] 1 2
vec_order(c(1, NA), na_value = "largest", direction = "desc")
#> [1] 2 1