Find and count unique values

vec_unique(): the unique values. Equivalent to unique().

vec_unique_loc(): the locations of the unique values.

vec_unique_count(): the number of unique values.

vec_unique(x)

vec_unique_loc(x)

vec_unique_count(x)

Arguments

x
A vector (including a data frame).

Value

vec_unique(): a vector the same type as x containing only unique values.

vec_unique_loc(): an integer vector, giving locations of unique values.

vec_unique_count(): an integer vector of length 1, giving the number of unique values.

Dependencies

vec_proxy_equal()

Missing values

In most cases, missing values are not considered to be equal, i.e. NA == NA is not TRUE. This behaviour would be unappealing here, so these functions consider all NAs to be equal. (Similarly, all NaN are also considered to be equal.)

See also

Examples

x <- rpois(100, 8)
vec_unique(x)
#>  [1] 13  7 10  9  4 11  5  8  6 12 19  3 15 14
vec_unique_loc(x)
#>  [1]  1  2  3  4  9 10 13 14 15 24 27 31 43 55
vec_unique_count(x)
#> [1] 14

# `vec_unique()` returns values in the order that encounters them
# use sort = "location" to match to the result of `vec_count()`
head(vec_unique(x))
#> [1] 13  7 10  9  4 11
head(vec_count(x, sort = "location"))
#>   key count
#> 1  13     7
#> 2   7    15
#> 3  10    10
#> 4   9    16
#> 5   4     6
#> 6  11     9

# Normally missing values are not considered to be equal
NA == NA
#> [1] NA

# But they are for the purposes of considering uniqueness
vec_unique(c(NA, NA, NA, NA, 1, 2, 1))
#> [1] NA  1  2